requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Deprecated==1.2.5
idna==2.8
PyGithub==1.43.7
PyJWT==1.7.1
requests==2.21.0
urllib3==1.24.2
wrapt==1.11.1
